Adds Dropwizard example
This is the first example to not use Brave configuration directly.

This uses https://github.com/smoketurner/dropwizard-zipkin to manage
tracer configuration.

Similar to the Spring Boot example, this makes a single file for both
examples: `brave.example.Frontend` and `brave.example.Backend` even if
it is more common in practice to split up things into many files.

This helps make the examples easier to maintain at the cost of being
slightly less idiomatic.

cc @jplock as this should eventually be linked to the dropwizard-zipkin
repository. Note this also makes a docker image!