Publishes to GitHub Container Registry instead of Docker Hub (#66)
As mentioned in https://github.com/openzipkin/docker-java/pull/33, we
are no longer pushing images to Docker Hub that are used in CI loops.

This prevents us and others from hitting the 100 or 200 Docker pull
limits applying on November 1st. While Zipkin images will continue to
push to DockerHub example images like these won't.

This implies changing your files like below:
```diff
-    image: openzipkin/example-brave:${PROJECT:-armeria}
+    image: ghcr.io/openzipkin/brave-example:${PROJECT:-armeria}
```

It not may seem intuitive to use Travis to push the images, but it is
quicker for me, and easy enough for someone else to refactor into a
GitHub action later. There are only a few more days before the new
limits apply and I have a lot of other work to do including unrelated
to Zipkin.